ADD: INOUT-51 AT http only cookie 적용 및 리액트 이슈 해결 및
- 리액트 http only cookie 허용 요청시 cors문제 해결을 위한
JwtAuthenticationFilter  OPTIONS 설정 추가
- AT, RT 생성 모듈화
- 회원 탈퇴 컨트롤러 구현
- handleUsernameNotFound 핸들러 구현
- 회원전용 컨트롤러 AT Http-Only Cookie사용되게 변경
- SecurityConfig options method 설정 추가
- SignInDto ValidationGroups SignInGroup으로 변경
- 시큐리티 적용된 컨트롤러들 테스트 코드 일부 제외 구현

+++ src/main/java/com/agora/debate/member/controller/MemberController.java
@@ -3,17 +3,19 @@ package com.agora.debate.member.controller;
 
 import com.agora.debate.member.dto.profile.MemberInfo;
 import com.agora.debate.member.entity.Member;
-import com.agora.debate.security.JwtUtils;
 import com.agora.debate.member.service.JwtBlacklistService;
 import com.agora.debate.member.service.MemberService;
+import com.agora.debate.security.utils.JwtCookieUtils;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-
+@Slf4j
 @RestController
 @RequestMapping("/members")
 @RequiredArgsConstructor
@@ -37,13 +39,14 @@ public class MemberController {
                 .build());
     }
 
-    /**
-     * TODO : AT도 httponly에서 지우기
-     */
     @PostMapping("/logout")
-    public ResponseEntity<?> logout(@RequestHeader("Authorization") String authorizationHeader,
-                                    @CookieValue(value = "refreshToken", required = false) String refreshToken) {
-        String accessToken = JwtUtils.extractAccessToken(authorizationHeader);
+    public ResponseEntity<?> logout(@CookieValue(value = "refreshToken", required = false) String refreshToken,
+                                    @CookieValue(value = "accessToken", required = false) String accessToken) {
+        if (accessToken == null && refreshToken == null) {
+            log.info("로그아웃 실패");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        log.info("로그아웃 진행");
 
         // 2. AT 블랙리스트 등록
         if (accessToken != null) {
@@ -54,24 +57,38 @@ public class MemberController {
         if (refreshToken != null) {
             jwtBlacklistService.addBlacklistToken(refreshToken,"logout");
         }
-        // 4. RT 쿠키 삭제 (Set-Cookie로 만료시키기)
-        ResponseCookie deleteCookie = ResponseCookie.from("refreshToken", "")
-                .maxAge(0)
-                .httpOnly(true)
-                .secure(true) // 개발환경이면 false
-                .path("/")
-                .build();
-
+        log.info("로그아웃 완료");
         return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, deleteCookie.toString())
+                .header(HttpHeaders.SET_COOKIE, JwtCookieUtils.deleteAccessTokenCookie().toString())
+                .header(HttpHeaders.SET_COOKIE, JwtCookieUtils.deleteRefreshTokenCookie().toString())
                 .body("로그아웃 완료");
     }
 
     @DeleteMapping
-    public ResponseEntity<?> deleteMember(@AuthenticationPrincipal Member member) {
+    public ResponseEntity<?> deleteMember(@AuthenticationPrincipal Member member,
+                                          @CookieValue(value = "refreshToken", required = false) String refreshToken,
+                                          @CookieValue(value = "accessToken", required = false) String accessToken) {
+        if (accessToken == null && refreshToken == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        // 2. AT 블랙리스트 등록
+        if (accessToken != null) {
+            jwtBlacklistService.addBlacklistToken(accessToken,"logout");
+        }
+
+        // 3. RT 블랙리스트 등록
+        if (refreshToken != null) {
+            jwtBlacklistService.addBlacklistToken(refreshToken,"logout");
+        }
+
         memberService.deleteMember(member);
 
-        return ResponseEntity.ok().body("회원 탈퇴가 완료되었습니다.");
+        return ResponseEntity
+                .ok()
+                .header(HttpHeaders.SET_COOKIE, JwtCookieUtils.deleteAccessTokenCookie().toString())
+                .header(HttpHeaders.SET_COOKIE, JwtCookieUtils.deleteRefreshTokenCookie().toString())
+                .body("회원 탈퇴가 완료되었습니다.");
     }
 
 }
